refactor: streamline dashboard page
- corrected broken and relative import paths for consistency
- deleted obsolete React components no longer in use
- improved overall dashboard maintainability and structure
